Update source/getstarted/dexterity.rst
added ./bin/ before virtualenv to make sure it works and using the virtualenv created

added note as their are a few questions asked during the dexterity product egg creation not mentioned.

added note to accept default values on add a content-type section

I am removing the complete fix section as fixing cfg files should not be done by each and every user. I am updating the the actual cfg files instead. Also it is the plone.cfg and not buildout.cfg that should be changed.